Parse fetch responses one line at a time
In the previous implementation, the whole fetch response was put in one
string but the string was then too big to be parsed.

Now for every incoming line for the server, we check if it is a fetch
answer. If it is a fetch answer, we parse it directly and parse the
following lines accordingly until the fetch response is complete.